Rename tileset folder and tab manager
