+++ game_core/editor/__init__.py
@@ -4,5 +4,5 @@ __all__ = ["Sidebar", "Canvas", "TabManager", "TilesetTabManager"]
 
 from .sidebar import Sidebar
 from .canvas import Canvas
-from .tab_manager import TabManager
+from .sidebar_tab_manager import TabManager
 from .tileset_tab.tileset_tab_manager import TilesetTabManager
