+++ game_core/editor/sidebar_tab_manager.py
