+++ game_core/editor/tileset_tab/tileset_components/show_tileset.py
